feat 파라미터 타입 변경

+++ sosik-intake-service/src/main/java/com/example/sosikintakeservice/service/redis/RedisIntakeService.java
@@ -41,7 +41,7 @@ public class RedisIntakeService {
         return prefix + KEY_PREFIX + ":" + memberId + ":" + period;
     }
 
-    public void save(String selectCondition, Long memberId, Long foodId, int period, Integer value) {
+    public void save(String selectCondition, Long memberId, Long foodId, int period, Double value) {
 
         String key = getKey(selectCondition, memberId, period);
